Document the use of the provided scope with Maven
This is consistent with the Gradle documentation using the compileOnly
configuration, and avoids the inclusion of neo4j-sproc-compiler and its
dependencies when using the shade plugin to package the procedures jar.

The optional attribute is kept to avoid pulling in the compiler when using
the procedures jar as a dependency, e.g. for an embedded Neo4j instance.