Explicitly reject all invalid types
The default behavior of the Simple*Visitors is to return null for any
case that has not been overridden.
When, for example, a procedure return type defines a field with
an array type, the RecordFieldTypeVisitor returns `null` instead of
`false`, leading to an unhelpful `NullPointerException` in
https://github.com/fbiville/neo4j-sproc-compiler/blob/34a718cc6793a2c96732cdc14ccf22d7f0652b4d/processor/src/main/java/net/biville/florent/sproccompiler/visitors/RecordTypeVisitor.java#L74
instead of the more helpful error message that the
target type is not supported.

With this commit, the validating visitors explicitly return `false`
instead of `null` on types that aren't covered directly.